+++ bindings/matrix-sdk-ffi/src/authentication_service.rs
@@ -9,6 +9,7 @@ use url::Url;
 use zeroize::Zeroize;
 
 use super::{client::Client, client_builder::ClientBuilder, RUNTIME};
+use crate::error::ClientError;
 
 pub struct AuthenticationService {
     base_path: String,
@@ -144,7 +145,7 @@ impl AuthenticationService {
             }
         }
 
-        let client = builder.build().or_else(|e| {
+        let client = builder.build_inner().or_else(|e| {
             if !server_name_or_homeserver_url.starts_with("http://")
                 && !server_name_or_homeserver_url.starts_with("https://")
             {
@@ -153,7 +154,7 @@ impl AuthenticationService {
             // When discovery fails, fallback to the homeserver URL if supplied.
             let mut builder = Arc::new(ClientBuilder::new()).base_path(self.base_path.clone());
             builder = builder.homeserver_url(server_name_or_homeserver_url);
-            builder.build()
+            builder.build_inner()
         })?;
 
         let details = RUNTIME.block_on(self.details_from_client(&client))?;
@@ -186,7 +187,9 @@ impl AuthenticationService {
 
         // Login and ask the server for the full user ID as this could be different from
         // the username that was entered.
-        client.login(username, password, initial_device_name, device_id)?;
+        client.login(username, password, initial_device_name, device_id).map_err(|e| match e {
+            ClientError::Generic { msg } => AuthenticationError::Generic { message: msg },
+        })?;
         let whoami = client.whoami()?;
 
         // Create a new client to setup the store path now the user ID is known.
@@ -208,7 +211,7 @@ impl AuthenticationService {
             .homeserver_url(homeserver_url)
             .sliding_sync_proxy(sliding_sync_proxy)
             .username(whoami.user_id.to_string())
-            .build()?;
+            .build_inner()?;
 
         // Restore the client using the session from the login request.
         client.restore_session_inner(session)?;
@@ -260,7 +263,7 @@ impl AuthenticationService {
             .passphrase(self.passphrase.clone())
             .homeserver_url(homeserver_url)
             .username(whoami.user_id.to_string())
-            .build()?;
+            .build_inner()?;
 
         // Restore the client using the session.
         client.restore_session_inner(session)?;