+++ crates/matrix-sdk/CHANGELOG.md
@@ -1,3 +1,9 @@
+# unreleased
+
+- `Common::members` and `Common::members_no_sync` take a `RoomMemberships` to be able to filter the
+  results by any membership state.
+  - `Common::active_members(_no_sync)` and `Common::joined_members(_no_sync)` are deprecated.
+
 # 0.6.2
 
 - Fix the access token being printed in tracing span fields.